IGNITE-21602 / move types to the top level
